caddy-wrapper.sh: explicitly handle -h|--help
Under some (as yet unknown conditions) the script goes a little wonky
and thinks `-h` is the name of a package I would like to add via `caddy
add-package`, which is clearly not true.

I haven't experienced this problem with other flags; it may be prudent
to do a better argparse in general.