snap: hooks: configure: ensure service is enabled
If a config is set for the caddy snap the server service should be
enabled as the intention of the config option is expressly for this
purpose :)

Note that this check (and other config-file-related things) should fall
after the [ -n "$_config" ] check, as the server should only ever be
enabled if a config is passed from snap configuration.

It's possible that in the future this behavior may change (specifying a
configuration file rather than encoding everything into the config key's
value), but for now this is the behavior.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>